Initial import of lutok
Lutok is a lightweight C++ API for Lua.  See the README file for a more
detailed description.

The code imported in this revision comes from the utils::lua module of
the Kyua project (http://code.google.com/p/kyua/), and aims to replace
such module in the short/medium term.

The original code has been modified to allow it to build without any
dependencies from Kyua and to rename it from utils::lua to lutok.
However, I have attempted to keep all changes to a minimum; improvements
can be done iteratively on top of this codebase to keep a record of what
has changed.

This change was r2 in Subversion.