Split the wrap module into state and stack_cleaner modules
The wrap module used to contain two very different classes: state, which
is the lightweight wrapper over the Lua C API, and stack_cleaner, which
is our own creation to deal with inconsistencies in the Lua stack.  These
classes belong in different modules, so this is exactly what this change
does.

This change was r6 in Subversion.