Replace push_c_{closure,function} with push_cxx_{closure,function}
The new push_cxx_{closure,function} methods of lutok::state completely
hide all the details of how to wrap a C++ function into a C function
that can safely be called from within Lua.

The upside of this is that we can finally avoid exposing lua_CFunction
to code including state.hpp (with the goal of later killing the
inclusion of lua.hpp), and also that caller code is much simplified
because there is no need to use the wrap_cxx_function template.

The downside is the obvious: we incur an extra run-time cost in doing
the above due to the way we store state to recover the C++ function data
from the C context.  This is not a big deal though, because it goes in
line with the rest of the design of the library.

I have ideas on how to cleanly provide push_c_{closure,function} again
but this will come separately once state.hpp is completely C-free.

This change was r8 in Subversion.