Move debug-related functions to a new debug module
The new debug module provides a wrapper over the lua_Debug C data type
so that this does not have to be exposed to C++ code -- not even in the
form of a typedef.

As a result of this change, the get_info and get_stack functions have
been moved from the state class into the new debug class.  I'm not too
fond of the current API of the debug module, but we can change this
later.  The goal at the moment is to get rid of the lua.hpp inclusion
from header files.

This change was r9 in Subversion.