Introduce the c_gate module (and get rid of lua.hpp from state.hpp)
This change makes the state.hpp module completely C-clean by moving the
functions that provide direct access to the Lua C state to their own
module.

This new module, c_gate, is explicitly named to represent that it provides
raw access to the Lua C API and that problems may arise due to this.

This change was r10 in Subversion.