Get rid of the lua.hpp dependency from the exceptions module
Make the api_error::from_stack method defined in the exceptions module
take a lutok::state instead of a lua_State.  This finally allows us to
completely drop the exposing of lua.hpp from our public headers (with
the exception of c_gate.hpp, but that's expected).

This change was r12 in Subversion.