Fix installation under Ubuntu 10.04.1 LTS
The installation of the html documentation using pax was failing in
Ubuntu 10.04.1 because pax was not creating the necessary target
directories (as the manual page promises).

Fix this by making the installation of the html documents depend on
the installation of the plain-text documents so that the target
directory gets created beforehand.

This change was r22 in Subversion.