Include the header file of the module under test before everything else
This little trick allows us to validate that the header file is
self-contained because no other includes have happened yet.

Because there is one (and only one) test program per test module, this
means that we are effectively ensuring that all header files in the
source tree are self-contained.

This change was r23 in Subversion.