Put pregenerated HTML documentation into distfiles
Make the install rules place pregenerated HTML documentation into the
distfiles, either if this documentation has been created as part of the
build or if it was shipped within the distfile.

As a side effect, this needs the addition of special uninstall hooks to
cope with this and also a change in the Doxyfile to prevent generating
large long names that gnutar cannot handle.

As usual, these Makefile rules are tricky, but they appear to work fine
in Mac OS X and NetBSD with both gmake and bmake.

This change was r26 in Subversion.