Add some examples
These examples will be linked from the wiki site to showcase how Lutok
works.  They are part of the source tree and they come with a little
test program to ensure that they always work and don't get out of sync
with changes to the library.

This change was r27 in Subversion.