Fix the examples tests in NetBSD
Ensure that the examples can find the liblutok library by defining
LD_LIBRARY_PATH during testing.  This is not an issue when using pkgsrc
because our own lutok.pc has been fixed to include an rpath... but it
is a problem when building directly from the source tree.

Additionally, fix the call to pkg-config in the examples Makefile so
that we get all the flags in the lutok.pc file (in particular, the
rpath to liblua).  It turns out that passing --libs-only-L and
--libs-only-other at the same time to pkg-config results in one of the
flags being ignored.

This change was r30 in Subversion.