Get rid of api-docs/html/installdox
Doxygen generates an installdox script inside the HTML output directory.
This script relies in perl, which triggers some misleading warnings in
pkgsrc.  As we don't need this script, just delete it to prevent it from
ending in the distribution and in the installed system.

This change was r31 in Subversion.