Support Lua 5.2 when accessing the globals table
Lua 5.2 has changed the way to access the globals table: instead of having
a pseudo-index, it needs to be referenced via a value in the registry.

Implement the get_global_table() to generalize access to the globals table
in a way that works with both Lua 5.1 and 5.2.

While doing this, I have to kill the globals_index symbol as it cannot be
represented in Lua 5.2.  Instead, expose a registry_index symbol.