Add initial GitHub Actions support for lutok
This proposed GitHub Actions workflow file runs tests on MacOS and
Ubuntu Linux, similar to what is being proposed in
https://github.com/freebsd/atf/pull/93 .

The delta between this change and the one being made in ATF mostly
relates to the additional package dependencies, e.g., Lua, and
build permutations, e.g., with and without ATF.

Kyua is built from scratch because of issues with dependent packages and
runtime linker collisions on both MacOS and Ubuntu. The binary package
versions are built against a different version of Lutok, ergo they can
have runtime linker collisions.

Signed-off-by: Enji Cooper <ngie@FreeBSD.org>