Port over CI work from freebsd/atf
This change takes the CI work from the atf project that's been incubated
and iterated on over the past several weeks/months.

This work enables CI workflows with FreeBSD on Cirrus CI with
working/supported image families and with macOS/Ubuntu Linux on GitHub
Actions.

Much of this work was done in collaboration with and building on work
started by @kevemueller and @mchoo7 in that respective repo.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>